Remote page - model and docs updated
