+++ core/src/main/java/com/adobe/aem/spa/project/core/internal/impl/PageImpl.java
@@ -26,7 +26,6 @@ import org.apache.sling.models.annotations.Via;
 import org.apache.sling.models.annotations.injectorspecific.InjectionStrategy;
 import org.apache.sling.models.annotations.injectorspecific.ScriptVariable;
 import org.apache.sling.models.annotations.injectorspecific.Self;
-import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;
 import org.apache.sling.models.annotations.via.ResourceSuperType;
 import org.apache.sling.models.factory.ModelFactory;
 import org.jetbrains.annotations.NotNull;
@@ -68,16 +67,12 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
  * ({@link PageImpl#getExportedItemsOrder()} to :itemsOrder</li>
  * </ul>
  */
-@Model(
-    adaptables = SlingHttpServletRequest.class,
-    adapters = { Page.class, ContainerExporter.class },
-    resourceType = { PageImpl.RESOURCE_TYPE, PageImpl.RESOURCE_TYPE_REMOTE  }
-    )
+@Model(adaptables = SlingHttpServletRequest.class, adapters = { Page.class,
+    ContainerExporter.class }, resourceType = PageImpl.RESOURCE_TYPE)
 @Exporter(name = ExporterConstants.SLING_MODEL_EXPORTER_NAME, extensions = ExporterConstants.SLING_MODEL_EXTENSION)
 public class PageImpl implements Page {
 
     static final String RESOURCE_TYPE = "spa-project-core/components/page";
-    static final String RESOURCE_TYPE_REMOTE = "spa-project-core/components/remotepage";
 
     // Delegated to Page v1
     @ScriptVariable
@@ -106,9 +101,6 @@ public class PageImpl implements Page {
     @Self
     @Via(type = ResourceSuperType.class)
     private com.adobe.cq.wcm.core.components.models.Page delegate;
-
-    @ValueMapValue(injectionStrategy = InjectionStrategy.OPTIONAL)
-    private String remoteSPAUrl;
 
     /**
      * {@link Map} containing the page models with their corresponding paths (as keys)
@@ -255,12 +247,6 @@ public class PageImpl implements Page {
         return delegate.getExportedType();
     }
 
-    // Delegated to Page v1
-    @Override
-    public Set<String> getComponentsResourceTypes() {
-       return delegate.getComponentsResourceTypes();
-    }
-
     // Delegated to Page v2
     @Nullable
     @Override
@@ -321,10 +307,10 @@ public class PageImpl implements Page {
         return delegate.hasCloudconfigSupport();
     }
 
-    // Delegated to Page v2
+    // Delegated to Page v2 
     @NotNull
     @Override
-    public String getRemoteSPAUrl() {
-        return remoteSPAUrl;
+    public Set<String> getComponentsResourceTypes() {
+        return delegate.getComponentsResourceTypes();
     }
 }
