+++ ui.apps/src/main/content/jcr_root/apps/spa-project-core/components/remotepage/remotepage.html
@@ -11,7 +11,7 @@
 -->
 <!DOCTYPE html>
 <html
-  data-sly-use.page="com.adobe.aem.spa.project.core.models.Page"
+  data-sly-use.page="com.adobe.aem.spa.project.core.models.RemotePage"
   lang="${page.language}"
   data-sly-use.head="head.html"
   data-sly-use.footer="footer.html"
